requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.2
tavily-python==0.3.0
pydantic==2.5.0
gunicorn==21.2.0